+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### ENHANCEMENTS
+- Added securityContext configuration capability to the chart
 
 ## [0.2.0] - 2019-05-27
 ### ENHANCEMENTS